refactor: add reporters directory
