+++ packages/salesforcedx-utils-vscode/src/index.ts
@@ -39,7 +39,7 @@ export {
 } from './telemetry/telemetry';
 export * from './helpers';
 export { TraceFlags } from './helpers/traceFlags';
-export { AppInsights as TelemetryReporter } from './telemetry/appInsightsReporter';
+export { AppInsights as TelemetryReporter } from './telemetry/reporters/appInsightsReporter';
 export {
   hasRootWorkspace,
   getRootWorkspace,
