+++ packages/salesforcedx-utils-vscode/src/telemetry/reporters/telemetryLogger.ts
@@ -8,9 +8,9 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { Disposable, workspace } from 'vscode';
-import { getRootWorkspacePath } from '..';
-import { WorkspaceContextUtil } from '../context/workspaceContextUtil';
-import { TelemetryReporter } from './TelemetryReporterInterface';
+import { getRootWorkspacePath } from '../..';
+import { WorkspaceContextUtil } from '../../context/workspaceContextUtil';
+import { TelemetryReporter } from './telemetryReporter';
 
 export class TelemetryLogger extends Disposable implements TelemetryReporter {
   private toDispose: Disposable[] = [];
