+++ packages/salesforcedx-utils-vscode/src/telemetry/reporters/telemetryReporter.ts
