+++ packages/salesforcedx-utils-vscode/src/telemetry/telemetry.ts
@@ -13,10 +13,10 @@ import {
   SFDX_EXTENSION_PACK_NAME
 } from '../constants';
 import { SfdxSettingsService } from '../settings';
-import { AppInsights } from './appInsightsReporter';
+import { AppInsights } from './reporters/appInsightsReporter';
 import { disableCLITelemetry, isCLITelemetryAllowed } from './cliConfiguration';
-import { TelemetryLogger } from './telemetryLogger';
-import { TelemetryReporter } from './TelemetryReporterInterface';
+import { TelemetryLogger } from './reporters/telemetryLogger';
+import { TelemetryReporter } from './reporters/telemetryReporter';
 
 interface CommandMetric {
   extensionName: string;
